+++ core-testing/src/main/java/com/jraska/github/client/FakeModules.kt
@@ -1,5 +1,6 @@
 package com.jraska.github.client
 
+import com.jraska.github.client.android.FakeCoreAndroidModule
 import com.jraska.github.client.http.FakeHttpModule
 import dagger.Module
