+++ core-testing/src/main/java/com/jraska/github/client/Fakes.kt
@@ -1,7 +1,7 @@
 package com.jraska.github.client
 
 import com.jraska.github.client.analytics.EventAnalytics
-import com.jraska.github.client.logging.CrashReporter
+import com.jraska.github.client.android.RecordingNavigator
 import com.jraska.github.client.rx.AppSchedulers
 import io.reactivex.schedulers.Schedulers
