+++ feature/repo/build.gradle
@@ -45,4 +45,5 @@ dependencies {
   testImplementation 'junit:junit:4.13.2'
   testImplementation 'org.assertj:assertj-core:3.18.1'
   testImplementation project(':core-testing')
+  kaptTest rootProject.ext.daggerAnnotationProcessor
 }
